MaxCompatSupertree
1) MaxCompatSupertree - prints # of cols/chars removed, actual columns removed,
and newick string of inferred supertree

2) CharacterIntGraphFR & descendents use dynamic allocation

3) const X* to const X& as function input parameter where encountered

4) const X* to const X& as function return type where encountered

5) PIG & CIG vertices are back to having i#j names instead of i,j names.
This is because clique trees name their nodes with, say,
{i#j,k#l}
which is clear, vs.
{i,j,k,l}
which is all sorts of awful.

6) Graph(const Graph&) constructor added.

7) SetupAndRunHeuristic / Compatible are now static members of HeuristicRun.
SetupAndRunHeuristic is now New. Run stats are now written to strings instead
of written to cout (TriCHe will need to be updated).

8) IsChordal -> IsTriangulated

9) Triangulation::New(const Graph* G) checks if G is triangulated before
returning a triangulation

10) Added comments to headers